refactor: update blog posts slug
