+++ src/app/[slug]/posts/a-brief-history-of-javascript.mdx
