+++ src/app/[slug]/posts/animated-gradient.mdx
