+++ src/app/[slug]/posts/equality-of-values-in-javascript.mdx
