+++ src/app/[slug]/posts/games.mdx
