+++ src/app/[slug]/posts/gatsby-image-medium-zoom-effect.mdx
