+++ src/app/[slug]/posts/gatsby-location.mdx
