+++ src/app/[slug]/posts/gatsby-smooth-scrolling.mdx
