+++ src/app/[slug]/posts/how-to-add-custom-javascript-to-a-gatsby-site.mdx
