+++ src/app/[slug]/posts/how-to-hide-the-address-bar-in-a-progressive-web-application.mdx
