+++ src/app/[slug]/posts/jamstack-best-practices.mdx
