+++ src/app/[slug]/posts/javascript-objects-and-functions.mdx
