+++ src/app/[slug]/posts/javascript-primitive-values.mdx
