+++ src/app/[slug]/posts/javascript-properties.mdx
