+++ src/app/[slug]/posts/javascript-prototypes.mdx
