+++ src/app/[slug]/posts/javascript-values-and-variables.mdx
