+++ src/app/[slug]/posts/my-eslint-and-prettier-config.mdx
