+++ src/app/[slug]/posts/my-macbook-setup-for-web-development.mdx
