+++ src/app/[slug]/posts/pure-functions-in-javascript.mdx
