+++ src/app/[slug]/posts/react-components-vs-elements.mdx
