+++ src/app/[slug]/posts/react-events.mdx
