+++ src/app/[slug]/posts/react-rendering.mdx
