+++ src/app/[slug]/posts/setting-up-namecheap-domain-in-vercel.mdx
