+++ src/app/[slug]/posts/strict-mode.mdx
