+++ src/app/[slug]/posts/text-gradient.mdx
