+++ src/app/[slug]/posts/the-fundamentals.mdx
