+++ src/app/[slug]/posts/the-script-tag.mdx
