+++ src/app/[slug]/posts/understanding-the-lifecycle-of-gatsby.mdx
