+++ src/app/[slug]/posts/variable-declarations-var-let-const.mdx
