+++ src/app/[slug]/posts/what-is-jamstack.mdx
