+++ src/lib/mdx.ts
@@ -50,5 +50,5 @@ function getMDXData(dir: string) {
 }
 
 export function getBlogPosts() {
-  return getMDXData(path.join(process.cwd(), "src", "app", "blog", "posts"));
+  return getMDXData(path.join(process.cwd(), "src", "app", "[slug]", "posts"));
 }
